+++ initializers/table_importer.gd
@@ -19,67 +19,57 @@
 # *****************************************************************************
 class_name IVTableImporter
 
-# Reads external data tables (.tsv files) and adds results to:
-#    - IVGlobal.table_rows
-#    - IVGlobal.wiki_titles (if IVGlobal.wiki_enabled)
-#    - table containers in IVTableReader
+# Reads external data tables (.tsv files) and adds processed results to
+# IVGlobal dictionaries. These dictionaries are indexed as follows:
 #
-# Use IVTableReader to interact with imported table data! This object removes
-# itself from IVGlobal.program after table import. No other object should
-# reference it.
+#    tables[table_name][column_field][row_name or row_int] -> typed_value
+#    table_rows[row_name] -> row_int (row_name's are globally unique)
+#    table_types[table_name][column_field] -> Type string in table
+#    table_precisions[][][] indexed as tables w/ REAL fields only -> sig digits
+#    wiki_titles[row_name] -> title string for wiki target resolution
 #
-# ivoyager/data/solar_system/*.tsv table construction:
-#  Data_Type (required!): X, BOOL, INT, FLOAT, STRING, ENUM, TABLE_ROW, BODY
-#    See tables for examples; see IVTableReader for conversions.
-#  Default (optional; all types except X): If cell is blank, it will be
-#    replaced with this value.
-#  Units (optional; REAL only!). Reals will be converted from provided units
+# IVTableReader provides protected access and constructor methods.
+#
+# See data/solar_system/README.txt for table construction. In short:
+#
+#  Type (required!): X, BOOL, INT, FLOAT, STRING, TABLE_ROW or enum name.
+#    An enum must be present in static file referenced in IVGlobal.enums. 'X'
+#    type is really just BOOL where 'x' is true and blank cell is false.
+#  Default (optional; not expected for X): Use this value if blank cell.
+#  Units (optional; REAL only!): Reals will be converted from provided units
 #    symbol. The symbol must be present in IVUnits.MULTIPLIERS or FUNCTIONS or
 #    replacement dicts specified in IVGlobal.unit_multipliers, .unit_functions.
-#
-# False is represented internally as "", so bool(internal_value) will give
-# meaningful result for both BOOL and X type.
+
 
 const units := preload("res://ivoyager/static/units.gd")
+const utils := preload("res://ivoyager/static/utils.gd")
 const math := preload("res://ivoyager/static/math.gd")
 
 const DPRINT := false
-const DATA_TYPES := ["REAL", "BOOL", "X", "INT", "STRING", "BODY", "TABLE_ROW"] # & enum names
+const DATA_TYPES := ["REAL", "BOOL", "X", "INT", "STRING", "TABLE_ROW"] # & enum names
 
 # source files
 var _table_import: Dictionary = IVGlobal.table_import
 var _wiki_titles_import: Array = IVGlobal.wiki_titles_import
 # imported data
-var _table_data := {}
-var _table_fields := {}
-var _table_data_types := {}
-var _table_units := {}
-var _table_row_dicts := {}
+var _tables: Dictionary = IVGlobal.tables
 var _table_rows: Dictionary = IVGlobal.table_rows # IVGlobal shared
+var _table_types: Dictionary = IVGlobal.table_types # indexed [table_name][field]
+var _table_precisions: Dictionary = IVGlobal.table_precisions # as _tables for REAL fields
 var _wiki_titles: Dictionary = IVGlobal.wiki_titles # IVGlobal shared
-var _unique_register := {}
 
 # localization
 var _enable_wiki: bool = IVGlobal.enable_wiki
 var _enums: Script = IVGlobal.enums
 var _wiki: String = IVGlobal.wiki # wiki column header
+var _unit_multipliers: Dictionary = IVGlobal.unit_multipliers
+var _unit_functions: Dictionary = IVGlobal.unit_functions
 
-# current processing
-var _path: String
-var _data: Array
-var _fields: Dictionary
-var _rows: Dictionary
-var _data_types: Array
-var _units: Array
-var _defaults: Array
-var _line_array: Array
-var _row_name: String
-var _field: String
-var _data_type: String
-var _cell: String
+# data counting
 var _count_rows := 0
 var _count_cells := 0
 var _count_non_null := 0
+var _unique_register := {}
 
 
 func _init():
@@ -95,212 +85,245 @@ func _on_init() -> void:
 
 
 func _project_init() -> void:
-	var table_reader: IVTableReader = IVGlobal.program.TableReader
-	table_reader.init_tables(_table_data, _table_fields, _table_data_types, _table_units, _table_row_dicts)
-	IVGlobal.program.erase("TableImporter") # frees self
+	IVGlobal.program.erase("TableImporter2") # frees self
 
 
 func _import() -> void:
 	for table_name in _table_import:
-		_path = _table_import[table_name]
-		_data = []
-		_fields = {} # column index by field name
-		_data_types = []
-		_rows = {} # row index by name
-		_read_table()
-		# wiki_titles was populated on the fly (if IVGlobal.enable_wiki); but we
-		# save everything else to IVGlobal dicts below
-		_table_data[table_name] = _data
-		_table_fields[table_name] = _fields
-		for i in range(_data_types.size()):
-			if _data_types[i] == "X":
-				_data_types[i] = "BOOL"
-		_table_data_types[table_name] = _data_types
-		_table_units[table_name] = _units
-		_table_row_dicts[table_name] = _rows
-		for row_name in _rows:
-			assert(!_table_rows.has(row_name))
-			_table_rows[row_name] = _rows[row_name]
+		var path: String = _table_import[table_name]
+		_import_table(table_name, path)
+	_postprocess()
 	if !_enable_wiki:
 		return
-	_data = []
-	_fields = {}
-	_data_types = []
-	_rows = {}
 	for path in _wiki_titles_import:
-		_path = path
-		_read_table() # writes wiki_titles; we don't keep anything else
-		_data.clear()
-		_fields.clear()
-		_data_types.clear()
-		_rows.clear()
+		_import_wiki_titles(path)
 
 
-func _read_table() -> void:
-	assert(DPRINT and prints("Reading", _path) or true)
+func _import_wiki_titles(path: String) -> void:
+	assert(DPRINT and prints("Reading", path) or true)
 	var file := File.new()
-	if file.open(_path, file.READ) != OK:
-		assert(false, "Could not open file: " +  _path)
-	_units = []
-	_defaults = []
-	var have_fields := false
-	var reading_data := false
+	if file.open(path, file.READ) != OK:
+		assert(false, "Could not open file: " +  path)
+	var fields := {}
+	var reading_header := true
+	var reading_fields := true
 	var n_columns := 0
 	var line := file.get_line()
 	while !file.eof_reached():
-		var commenter := line.find("#")
-		if commenter != -1 and commenter < 4: # skip comment line
+		var comment_test := line.find("#")
+		if comment_test != -1 and comment_test < 4: # skip comment line
 			line = file.get_line()
 			continue
-		_line_array = line.split("\t")
-		if !reading_data:
-			if !have_fields: # always 1st line!
-				assert(_line_array[0] == "name", "1st field must be 'name'")
-				for field in _line_array:
-					if field == "Comments":
-						break
-					_fields[field] = n_columns
+		var line_array := line.split("\t") as Array
+		if reading_header:
+			if reading_fields: # always 1st line!
+				var cell00: String = line_array[0]
+				assert(cell00 == "name", "1st field must be 'name'")
+				for field in line_array:
+					fields[field] = n_columns
 					n_columns += 1
-				have_fields = true
-				assert(n_columns == _fields.size(), "Duplicate field (%s columns, %s unique fields) in %s" \
-						% [n_columns, _fields.size(), _path])
-			elif _line_array[0] == "Type":
-				_data_types = _line_array.duplicate()
-				_data_types[0] = "STRING" # always name field
-				_data_types.resize(n_columns) # truncate Comment column
-				assert(_data_types_test())
-			elif _line_array[0] == "Units":
-				_units = _line_array.duplicate()
-				_units[0] = ""
-				_units.resize(n_columns) # truncate Comment column
-				assert(_units_test())
-			elif _line_array[0] == "Default":
-				_defaults = _line_array.duplicate()
-				_defaults[0] = ""
-				_defaults.resize(n_columns) # truncate Comment column
+				reading_fields = false
+				assert(n_columns == fields.size(), "Duplicate field (%s columns, %s unique fields) in %s" \
+						% [n_columns, fields.size(), path])
+			elif line_array[0] == "Type": # REMOVE after conversion
+				pass
 			else:
-				assert(_data_types) # required
-				if !_units:
-					for _i in range(n_columns):
-						_units.append("")
-				if !_defaults:
-					for _i in range(n_columns):
-						_defaults.append("")
-				assert(_table_test(n_columns))
-				reading_data = true
-		if reading_data:
-			_read_data_line()
+				reading_header = false
+		if reading_header:
+			line = file.get_line()
+			continue
+		var row_name: String = line_array[0]
+		assert(row_name, "name cell is blank!")
+		for field in fields:
+			if _enable_wiki and field == _wiki:
+				_count_cells += 1
+				var column: int = fields[field]
+				var value: String = line_array[column]
+				_wiki_titles[row_name] = value
 		line = file.get_line()
 
 
-func _read_data_line() -> void:
-	var row := _data.size()
-	var row_data := []
-	row_data.resize(_fields.size()) # unfilled row_data are nulls
-	_row_name = _line_array[0]
-	assert(_row_name, "name cell is blank!")
-	assert(!_rows.has(_row_name), "name is already used in this or another table!")
-	_rows[_row_name] = row
-	for field in _fields:
-		_count_cells += 1
-		_field = field
-		var column: int = _fields[_field]
-		_cell = _line_array[column]
-		_data_type = _data_types[column]
-		if !_cell: # set to default
-			_cell = _defaults[column]
-		_process_cell_value()
-		if !_cell:
-			row_data[column] = ""
+func _import_table(table_name: String, path: String) -> void:
+	assert(table_name and path)
+	assert(DPRINT and prints("Reading", path) or true)
+	var file := File.new()
+	if file.open(path, file.READ) != OK:
+		assert(false, "Could not open file: " +  path)
+	
+	var row := 0
+	_tables[table_name] = {}
+	_table_types[table_name] = {}
+	_table_precisions[table_name] = {}
+	var fields := {}
+	var types: Array
+	var units: Array
+	var defaults: Array
+	var reading_header := true
+	var reading_fields := true
+	var n_columns := 0
+	var line := file.get_line()
+	var has_type := false
+	var has_row_names: bool
+	while !file.eof_reached():
+		var comment_test := line.find("#")
+		if comment_test != -1 and comment_test < 4: # skip comment line
+			line = file.get_line()
 			continue
-		assert(_cell_test())
-		row_data[column] = _cell
-		_unique_register[_cell] = null
-		if _enable_wiki and _field == _wiki:
-			_wiki_titles[_row_name] = _cell
-		_count_non_null += 1
-	_data.append(row_data)
-	_count_rows += 1
+		var line_array := line.split("\t") as Array
+		if reading_header:
+			if reading_fields: # always 1st line!
+				var cell00: String = line_array[0]
+				assert(cell00 == "name" or cell00 == "nil", "1st field must be 'name' or 'nil'")
+				has_row_names = cell00 == "name"
+				for field in line_array:
+					assert(field != "n_rows", "Disallowed field name")
+					if field == "Comments":
+						break
+					if field != "nil":
+						_tables[table_name][field] = {}
+					fields[field] = n_columns
+					n_columns += 1
+				reading_fields = false
+				assert(n_columns == fields.size(), "Duplicate field (%s columns, %s unique fields) in %s" \
+						% [n_columns, fields.size(), path])
+			# Type, Units & Default; only Type is required and must be before Default
+			elif line_array[0] == "Type":
+				types = line_array.duplicate()
+				types[0] = "STRING" # always name field (or nil)
+				types.resize(n_columns) # truncate Comment column
+#				assert(_data_types_test())
+				# REAL values have parallel dict w/ precisions
+				for field in fields:
+					var column: int = fields[field]
+					var type: String = types[column]
+					_table_types[table_name][field] = type
+					if type == "REAL":
+						_table_precisions[table_name][field] = {}
+				has_type = true
+			elif line_array[0] == "Units":
+				units = line_array.duplicate()
+				units[0] = ""
+				units.resize(n_columns) # truncate Comment column
+#				assert(_units_test())
+			elif line_array[0] == "Default":
+				assert(has_type)
+				defaults = line_array.duplicate()
+				defaults[0] = ""
+				defaults.resize(n_columns) # truncate Comment column
+			else:
+				# We are done reading header; must be at first data line.
+				assert(types) # required
+				if !units:
+					for _i in range(n_columns):
+						units.append("")
+				if !defaults:
+					for _i in range(n_columns):
+						defaults.append("")
+#				assert(_table_test(n_columns))
+				reading_header = false
+		# data line
+		if !reading_header:
+			_read_line(table_name, row, line_array, fields, types, units, defaults, has_row_names)
+			row += 1
+		line = file.get_line()
+	
+	_tables[table_name].n_rows = row
 
 
-func _process_cell_value() -> void:
-	if _cell.begins_with("\"") and _cell.ends_with("\""):
-		_cell = _cell.lstrip("\"").rstrip("\"")
-	_cell = _cell.lstrip("'")
-	_cell = _cell.lstrip("_")
-	if _data_type == "BOOL":
-		_cell = _cell.to_lower()
-	elif _data_type == "X":
-		if _cell == "x":
-			_cell = "true"
-	elif _data_type == "REAL":
-		_cell = _cell.replace("E", "e")
-	elif _data_type == "STRING":
-		_cell = _cell.c_unescape() # does not work for "\uXXXX"; Godot issue #38716
-		_cell = IVUtils.c_unescape_patch(_cell) # handles "\uXXXX"
+func _read_line(table_name: String, row: int, line_array: Array, fields: Dictionary,
+		types: Array, units: Array, defaults: Array, has_row_names: bool) -> void:
+	var row_name := ""
+	if has_row_names:
+		row_name = line_array[0]
+		assert(row_name, "name cell is blank!")
+		assert(!_table_rows.has(row_name))
+		_table_rows[row_name] = row
+
+	for field in fields:
+		if field == "nil":
+			continue
+		_count_cells += 1
+		
+		var column: int = fields[field]
+		var raw_value: String = line_array[column]
+		if !raw_value: # blank cell; set to default (which may be blank)
+			raw_value = defaults[column]
+		var type: String = types[column]
+		var unit: String = units[column]
+		_set_preprocessed(table_name, field, row, row_name, raw_value, type, unit)
 
 
-func _data_types_test() -> bool:
-	for data_type in _data_types:
-		assert(DATA_TYPES.has(data_type) or data_type in _enums, "Unknown Type: " + data_type)
-	return true
-
-
-func _units_test() -> bool:
-	for unit in _units:
-		if unit:
-			assert(units.is_valid_unit(unit, true, IVGlobal.unit_multipliers, IVGlobal.unit_functions),
-					"Unkown unit '" + unit + "' in " + _path)
-	return true
-
-
-func _table_test(n_columns: int) -> bool:
-	for column in range(n_columns):
-		var data_type: String = _data_types[column]
-		match data_type:
-			"BOOL":
-				assert(!_units[column], "Expected no Units for BOOL in" + _path)
-			"X":
-				assert(!_defaults[column], "Expected no Default for X type in" + _path)
-				assert(!_units[column], "Expected no Units for X type in" + _path)
-			"INT":
-				assert(!_units[column], "Expected no Units for INT in" + _path)
-			"REAL":
-				pass
-			"STRING", "TABLE_ROW", "BODY":
-				assert(!_units[column], "Expected no Units for " + data_type + " in " + _path)
-			_: # must be valid enum name
-				assert(!_units or !_units[column], "Expected no Units for " + data_type + " in " + _path)
-				assert(data_type in _enums, "Non-existent enum dict '" + data_type + "' in " + _path)
-	return true
-
-
-func _cell_test() -> bool:
-	# "" is always ok and not checked here; _process_cell_value() has already
-	# processed table values to our internal format (e.g., REAL "E" -> "e").
-	match _data_type:
-		"BOOL":
-			assert(_cell == "true" or _cell == "false" or _line_error("Expected BOOL"))
+func _set_preprocessed(table_name: String, field: String, row: int, row_name: String,
+			raw_value: String, type: String, unit: String):
+	# Convert BOOL, X, REAL, INT and enums, and process STRING; we'll convert
+	# TABLE_ROW in postprocess()
+	if raw_value.begins_with("\"") and raw_value.ends_with("\""):
+		raw_value = raw_value.lstrip("\"").rstrip("\"")
+	raw_value = raw_value.lstrip("'")
+	raw_value = raw_value.lstrip("_")
+	var value # untyped
+	match type:
 		"X":
-			assert(_cell == "true" or _line_error("X type must be x or blank cell"))
-		"INT":
-			assert(_cell.is_valid_integer() or _line_error("Expected INT"))
+			assert(!raw_value or raw_value == "x")
+			value = raw_value == "x"
+		"BOOL":
+			assert(raw_value.matchn("false") or raw_value.matchn("true"))
+			value = raw_value.matchn("true")
+		"STRING":
+			value = raw_value.c_unescape() # does not work for "\uXXXX"; Godot issue #38716
+			value = utils.c_unescape_patch(value) # handles "\uXXXX"
 		"REAL":
-			var real := _cell.lstrip("~")
-			assert(real == "?" or real.is_valid_float() or _line_error("Expected REAL"))
-		"STRING", "TABLE_ROW", "BODY":
-			pass
-		_: # must be valid enum name
-			assert(_enums[_data_type].has(_cell) or _line_error("Non-existent enum value " + _cell))
-	return true
+			# we determine precision here
+			var precision := -1
+			if !raw_value:
+				value = NAN
+			elif raw_value == "?":
+				value = INF
+			else:
+				raw_value = raw_value.replace("E", "e")
+				if raw_value.begins_with("~"):
+					precision = 1
+					raw_value = raw_value.lstrip("~")
+				else:
+					precision = utils.get_real_str_precision(raw_value)
+				var real := float(raw_value)
+				if unit:
+					real = units.convert_quantity(real, unit, true, true,
+							_unit_multipliers, _unit_functions)
+				value = float(real)
+			_table_precisions[table_name][field][row] = precision
+			if row_name:
+				_table_precisions[table_name][field][row_name] = precision
+		"INT":
+			value = int(raw_value) if raw_value else -1
+		"TABLE_ROW": # we'll get int in _postprocess()
+			value = raw_value
+		_: # must be blank or a valid enum name
+			if !raw_value:
+				value = -1
+			else:
+				var enum_dict: Dictionary = _enums[type]
+				value = enum_dict[raw_value]
+	# set value
+	_tables[table_name][field][row] = value
+	if row_name:
+		_tables[table_name][field][row_name] = value
+	if _enable_wiki and field == _wiki:
+		_wiki_titles[row_name] = value
 
 
-func _line_error(msg := "") -> bool:
-	print("ERROR in _read_data_line...")
-	if msg:
-		print(msg)
-	print("cell value: ", _cell)
-	print("row name   : ", _row_name)
-	print("field     : ", _field)
-	print(_path)
-	return false
+func _postprocess() -> void:
+	# convert type TABLE_ROW to ints
+	for table_name in _tables:
+		for field in _tables[table_name]:
+			if field == "n_rows":
+				continue
+			var type: String = _table_types[table_name][field]
+			if type == "TABLE_ROW":
+				for row_key in _tables[table_name][field]:
+					var raw_value: String = _tables[table_name][field][row_key]
+					if raw_value:
+						_tables[table_name][field][row_key] = _table_rows[raw_value]
+					else:
+						_tables[table_name][field][row_key] = -1
+