+++ CHANGELOG.md
@@ -15,11 +15,13 @@ Requires non-Git-tracked **ivoyager_assets-0.0.10**; find in [ivoyager releases]
 ### Changes
 * Redesigned the core selection object (now IVSelection) to be more easily extensible. It's a wrapper object that can hold anything that might be selected (in core ivoyager we only select Body instances). It is now dynamically generated.
 * Save/load system made more intuitive with new object persist const 'PERSIST_MODE' with values NO_PERSIST, PERSIST_PROPERTIES_ONLY and PERSIST_PROCEDURAL.
+* Overhauled table import system to allow quick, direct access of typed table data via IVGlobal dictionaries.
 
 ### API-breaking changes
 * Replaced class IVSelectionItem with IVSelection.
 * Changed function names in IVBodyRegistry for selection related actions.
-* Old persist const 'PERSIST_AS_PROCEDURAL_OBJECT: bool' replaced with 'PERSIST_MODE' with enum values (see above). 
+* Old persist const 'PERSIST_AS_PROCEDURAL_OBJECT: bool' replaced with 'PERSIST_MODE' with enum values (see above).
+* Removed Type 'BODY' from table import system.
 
 ### Bug fixes
 * Widgets fixed to work when GUI is added after solar system build.